[refactor] use pygit2 to parse diff in CI
revert some doc changes
