+++ docs/API-Reference/cpp_linter.git.rst
@@ -1,0 +1,5 @@
+``git`` module
+==============
+
+.. automodule:: cpp_linter.git
+    :members:
