--- docs/API-Reference/cpp_linter.git_lib.rst
@@ -1,5 +1,0 @@
-``git_lib`` module
-====================
-
-.. automodule:: cpp_linter.git_lib
-    :members:
