--- docs/API-Reference/cpp_linter.git_parse.rst
@@ -1,5 +1,0 @@
-``git_parse`` module
-====================
-
-.. automodule:: cpp_linter.git_parse
-    :members:
